requirements.txt
opencv-python
numpy
pyautogui
pystray
pillow
keyboard